增加位置和类型编辑

+++ ruoyi-billiard/src/main/java/com/ruoyi/billiard/service/IDeskPriceService.java
@@ -60,5 +60,5 @@ public interface IDeskPriceService
      */
     public int deleteDeskPriceByDeskPriceId(Long deskPriceId);
 
-    BigDecimal queryPriceByType(Long storeId, Integer deskType);
+    BigDecimal queryPriceByType(Long storeId, Long deskType);
 }
